Rename py-vehicle-compiler to py-vehicle-cli
