+++ py-vehicle-cli/.gitignore
@@ -1,9 +1,9 @@
 # Specific to this project
 /build/
-/src/vehicle_compiler/binding.py
-/src/vehicle_compiler/binding_wrap.c
-/src/vehicle_compiler/_binding*.so
-/src/vehicle_compiler/_binding*.pyd
+/src/vehicle_cli/binding.py
+/src/vehicle_cli/binding_wrap.c
+/src/vehicle_cli/_binding*.so
+/src/vehicle_cli/_binding*.pyd
 /wheelhouse/
 /wheelhouselocal/
