+++ py-vehicle-cli/MANIFEST.in
