+++ py-vehicle-cli/requirements-ci.txt
