+++ py-vehicle-cli/src/vehicle_cli/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+import vehicle_cli
+
+if __name__ == "__main__":
+    vehicle_cli.main(sys.argv[1:])
