+++ py-vehicle-cli/src/vehicle_cli/_binding.pyi
