+++ py-vehicle-cli/tests/test_vehicle_cli.py
@@ -1,0 +1,4 @@
+def test_main():
+    import vehicle_cli
+
+    vehicle_cli.main(["--version"])
