+++ py-vehicle-cli/vehicle-python-binding.cabal
@@ -40,9 +40,9 @@ foreign-library _binding
 
   if os(windows)
     options:      standalone
-    mod-def-file: src/vehicle_compiler/binding.def
+    mod-def-file: src/vehicle_cli/binding.def
 
-  c-sources:        src/vehicle_compiler/binding_wrap.c
+  c-sources:        src/vehicle_cli/binding_wrap.c
   hs-source-dirs:   src
   other-modules:    VehiclePythonBinding
   build-depends:
