--- py-vehicle-compiler/src/vehicle_compiler/__main__.py
@@ -1,6 +1,0 @@
-import sys
-
-import vehicle_compiler
-
-if __name__ == "__main__":
-    vehicle_compiler.main(sys.argv[1:])
