--- py-vehicle-compiler/tests/test_vehicle_compiler.py
@@ -1,4 +1,0 @@
-def test_main():
-    import vehicle_compiler
-
-    vehicle_compiler.main(["--version"])
